handwritten-text-recognition/requirements.txt
opencv-python>=4.8.0
easyocr>=1.7.0
transformers>=4.35.0
torch>=2.6.0
torchvision>=0.17.0
torchaudio>=2.6.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-Levenshtein>=0.21.0
matplotlib>=3.8.0
tqdm>=4.66.0
symspellpy>=6.7.1
wordfreq>=3.1.0

# NOTE: PyTorch installation on Windows requires using:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
